refactor: Move solana msg verification helpers to its own file.
