+++ ampd/src/solana/mod.rs
@@ -1,4 +1,4 @@
 pub mod json_rpc;
 pub mod ws_verifier;
-pub mod verifier;
+pub mod msg_verifier;
